requirements.txt
pytest==8.3.3
pytest-mock==3.14.0
beautifulsoup4==4.12.3
request==2.32.3
